move get_images and collate_batch to utils
